fix: Remove TOKEN_NOT reference from unary() - not defined yet
TOKEN_NOT isn't in the lexer yet, so removed it from unary().
Only supporting unary minus for now, which is all we need for
negative numbers in function arguments.

The unary() function now only handles:
- Unary minus: -5, -x

TOKEN_NOT can be added in a future update when we add
logical not operator to the lexer.